Devorch 3: R0 baseline, R1 Rust workspace, R2 vertical slice
Executes waves R0-R2 of the staged Go-to-Rust migration specified in
docs/upgrade/.

R0 - measure and freeze:
- Captured the Go baseline: build, vet, test, binary sizes, daemon idle
  RSS, endpoint latency. Evidence in docs/rust-migration/.
- Found three defects that go vet was silently failing on:
  - internal/orchestrator/brain.go was corrupt (duplicate package line,
    remainder collapsed onto one line with reversed line order, wrong
    module path). Never compiled, never imported. Not mechanically
    recoverable; whole package quarantined to archive/broken/.
  - internal/cli/unified.go printed an int64 with %.1f, which failed the
    entire internal/cli test binary because go test runs vet.
  - internal/session/manager.go copied a Session by value while holding
    its embedded mutex. Fixed with a lock-free snapshot type; on-disk
    JSON shape unchanged.
- Recorded that the tree has 359 Go files across 97 packages but only 3
  test files, so R3 parity tests must be written rather than ported.

R1 - Rust workspace skeleton:
- Eight crates plus the CLI and xtask. Full gate green, 78 tests.
- The Rust core uses a separate database file so two migration systems
  never share one file. Per-table disposition of all 22 Go migrations is
  in docs/rust-migration/sql-schema-map.md.
- ProcessSpec never builds a shell string and can replace the child
  environment outright, which is what makes a secret broker implementable.

R2 - vertical slice parity:
- devorch workspace create/list/remove/inventory, agent inspect, doctor.
- Change inventory runs all four git commands, so untracked files count
  toward candidate scope. Reproduced through the real CLI: the candidate
  that adds an unrequested NOTES.md is now distinguishable from one that
  does not, which git diff --numstat alone could not do.
- All four agent CLIs inspected by execution, not assumption. Every one
  supports structured headless output, so PTY is the fallback for all
  four rather than the integration path for any.

+++ internal/cli/unified.go
@@ -4782,7 +4782,7 @@ func (c *UnifiedCLI) showHardwareStatus() error {
 	memUsage, err := getMemoryUsage(profile)
 	if err == nil {
 		memGB := float64(profile.MemTotalMB) / 1024
-		memUsedGB := memUsage.UsedMB / 1024
+		memUsedGB := float64(memUsage.UsedMB) / 1024
 		memPercent := float64(memUsage.UsedMB) / float64(profile.MemTotalMB) * 100
 		fmt.Printf("  Memory Usage: %.1fGB/%.1fGB (%.1f%%)\n", memUsedGB, memGB, memPercent)
 	} else {
